+++ Scripts/download_frameworks.sh
@@ -14,8 +14,9 @@ fi
 echo "✅ Valid Xcode Project Directory"
 
 # fetch the version from the file Sources/ReclaimInAppSdk/Resources/ReclaimInAppSdk.version from git source
-RECLAIM_SDK_VERSION="0.1.2"
+RECLAIM_SDK_VERSION="0.2.0"
 DOWNLOAD_URL="https://reclaim-inapp-sdk.s3.ap-south-1.amazonaws.com/ios/${RECLAIM_SDK_VERSION}/ReclaimXCFrameworks.tar.gz"
+# DOWNLOAD_URL="http://localhost:8000/ios/${RECLAIM_SDK_VERSION}/ReclaimXCFrameworks.tar.gz"
 TEMP_ARCHIVE="ReclaimXCFrameworks.tar.gz"
 TARGET_DIR="ReclaimXCFrameworks"