+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.2.0
+
+- Updated breaking changes in the use of the module apis
+- Added use of capability access token
+- Add new way to fetch provider information from the host app
+- Add host notifications for session identity updates
+
 ## 0.1.4
 
 - Documentation updates.